git-svn-id: http://springhead.info/spr2/Springhead2/trunk@4363 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Creature/CRNeckController.cpp
@@ -30,7 +30,7 @@ void CRNeckController::Step(){
 	if (csHead) {
 		PHSolidIf*    soHead = csHead->GetPHSolid();
 		PHIKOriCtlIf* cpHead = NULL;
-		for (size_t i=0; i<csHead->NIKControlPoints(); ++i) {
+		for (int i=0; i<csHead->NIKControlPoints(); ++i) {
 			if (!cpHead) {
 				cpHead = csHead->GetIKControlPoint(i)->Cast();
 			} else {
